requirements.txt
# Telegram Libraries
telethon>=1.34.0
python-telegram-bot>=20.7

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Client for Groq API
aiohttp>=3.10.0
httpx>=0.25.0

# Async utilities
asyncio-throttle>=1.0.2
aiofiles>=23.2.0
